requirements.txt
flask
snowflake-connector-python
